requirements.txt
streamlit
sentence-transformers
chromadb
PyMuPDF
protobuf==3.20.3
torch
pysqlite3-binary
uuid
